Adding rest docs for category and order item switching them to IT tests as well. Also entity generator order items id were refactored to hold 1 item for test purposes. Closes #45
